rename gfx1250 half/mxscale gemms
